requirements.txt
python-telegram-bot==20.7
openai>=1.30.0